+++ src/screenComponents/radarView.cpp
@@ -24,7 +24,7 @@ GuiRadarView::GuiRadarView(GuiContainer* owner, string id, TargetsContainer* tar
     long_range(false),
     show_ghost_dots(false),
     show_sectors(true),
-    show_terrain(false),
+    show_warp_layer(false),
     show_waypoints(false),
     show_target_projection(false),
     show_missile_tubes(false),
@@ -65,7 +65,7 @@ GuiRadarView::GuiRadarView(GuiContainer* owner, string id, float distance, Targe
     long_range(false),
     show_ghost_dots(false),
     show_sectors(true),
-    show_terrain(false),
+    show_warp_layer(false),
     show_waypoints(false),
     show_target_projection(false),
     show_missile_tubes(false),
@@ -150,8 +150,8 @@ void GuiRadarView::onDraw(sf::RenderTarget& window)
         drawRenderTexture(mask_texture, background_texture, sf::Color::White, sf::BlendMultiply);
     if (show_sectors)
         drawSectorGrid(background_texture);
-    if (show_terrain)
-        drawTerrain(background_texture);
+    if (show_warp_layer)
+        drawWarpLayer(background_texture);
     drawRangeIndicators(background_texture);
     if (show_target_projection)
         drawTargetProjections(background_texture);
@@ -918,15 +918,20 @@ int GuiRadarView::calcGridScaleMagnitude(int scale_magnitude, int position)
     return scale_magnitude;
 }
 
-void GuiRadarView::drawTerrain(sf::RenderTarget &window){
-    if (gameGlobalInfo->terrain.defined){
-        sf::Sprite terrainMap;
-        textureManager.getTexture(gameGlobalInfo->terrain.textureName)->setSmooth(true);
-        textureManager.setTexture(terrainMap, gameGlobalInfo->terrain.textureName);
-        terrainMap.setPosition(worldToScreen(gameGlobalInfo->terrain.coordinates));
-        terrainMap.setScale(getScale() * gameGlobalInfo->terrain.scale, getScale()* gameGlobalInfo->terrain.scale);
-        terrainMap.setColor(sf::Color(255, 255, 255, 128)); // half transparent
-        window.draw(terrainMap);
+void GuiRadarView::drawWarpLayer(sf::RenderTarget &window){
+    if (gameGlobalInfo->use_warp_terrain)
+    {
+        if (gameGlobalInfo->layer[warp_layer].defined){
+            MapLayer &layer = gameGlobalInfo->layer[warp_layer];
+            sf::Sprite layerMap;
+            textureManager.getTexture(layer.textureName)->setSmooth(true);
+            textureManager.setTexture(layerMap, layer.textureName);
+            layerMap.setPosition(worldToScreen(layer.coordinates));
+            layerMap.setScale(getScale() * layer.scale, getScale()* layer.scale);
+            float transparency = std::max(0.0, std::min(25.0 / (getScale()*450.0), 128.0));
+            layerMap.setColor(sf::Color(255, 255, 255, transparency)); // half transparent
+            window.draw(layerMap);
+        }
     }
 }